[FEAT] Divided function & PenetrationCalculator.java done

+++ src/Defense.java
@@ -1,13 +1,13 @@
 public class Defense {
-    private int Armor; // 방어력
-    private int MagicResist; // 마법 저항력
+    private long Armor; // 방어력
+    private long MagicResist; // 마법 저항력
     private int Health; // 체력
 
-    public int getArmor() {
+    public long getArmor() {
         return Armor;
     }
 
-    public int getMagicResist() {
+    public long getMagicResist() {
         return MagicResist;
     }
